refactor: revert HTTP read timeout to 5 seconds
In an attempt to fix AAAA DNS resolution timeouts reported in #52,we had increased the HTTP read
timeout to 10 seconds. In moode-player/moode#701 it has been reported that this does not reliably
fix the issue, which only appears on a Raspberry Pi Zero 2W. As a workaround, add `options no-aaaa`
to `/etc/resolv.conf`.